Event update

+++ public/js/script.js
@@ -9,6 +9,7 @@ document.addEventListener("DOMContentLoaded", function () {
   const closePanel = document.getElementById("closePanel");
   const closePanelE = document.getElementById("closePanelE");
   const eventAddForm = document.getElementById("eventForm");
+  const categoryForm = document.getElementById("categoryForm");
   const cancelForm = document.getElementById("cancel");
   const imagePreview = document.getElementById("imagePreview");
   const banner = document.getElementById("event-banner");
@@ -45,6 +46,8 @@ document.addEventListener("DOMContentLoaded", function () {
     sidePanelE.style.marginLeft = "0%";
     sidePanelE.style.display = "none";
     sidePanelE.classList.remove("open");
+
+    categoryForm.reset();
   });
 
   closePanelE.addEventListener("click", () => {
@@ -53,6 +56,11 @@ document.addEventListener("DOMContentLoaded", function () {
     addCategoryBtn.style.display = "inline-block";
     addEventBtn.style.display = "inline-block";
     sidePanelE.classList.remove("open");
+
+    eventAddForm.reset();
+    imagePreview.src = "";
+    bannerBase64String = "";
+    eventLocation = {};
   });
 
   addEventBtn.addEventListener("click", () => {
@@ -291,7 +299,7 @@ function showCategoryFilters(categories) {
     "center-align",
     "white",
     "z-depth-1",
-    "active"
+    "activeCard"
   ); // Make "All" active by default
 
   allCard.innerHTML = `<h5>All</h5>`;
@@ -299,9 +307,9 @@ function showCategoryFilters(categories) {
   // Add event listener for the "All" category
   allCard.addEventListener("click", function () {
     document.querySelectorAll(".option-card").forEach((card) => {
-      card.classList.remove("active");
+      card.classList.remove("activeCard");
     });
-    allCard.classList.add("active");
+    allCard.classList.add("activeCard");
     getEvents();
   });
 
@@ -315,10 +323,10 @@ function showCategoryFilters(categories) {
 
     card.addEventListener("click", function () {
       document.querySelectorAll(".option-card").forEach((card) => {
-        card.classList.remove("active");
+        card.classList.remove("activeCard");
       });
 
-      card.classList.add("active");
+      card.classList.add("activeCard");
       getEventsByCategory(category._id);
     });
     cardList.appendChild(card);
@@ -353,6 +361,7 @@ async function getEventsByCategory(category) {
 }
 
 let markers = [];
+let currentInfoWindow = null;
 
 function setEventCardContainer(events) {
   const cardContainer = document.getElementById("event-container");
@@ -428,7 +437,11 @@ function setEventCardContainer(events) {
     });
 
     marker.addListener("click", function () {
+      if (currentInfoWindow) {
+        currentInfoWindow.close();
+      }
       infoWindow.open(map, marker);
+      currentInfoWindow = infoWindow;
     });
 
     hideSpinner();
@@ -471,7 +484,7 @@ function viewEvent(event) {
         event.organizerName
       } - ${event.organizerContact}</p>
       <p id="modal-location" class="modal-info">
-      <i class="material-icons">location_city</i>${
+      <i class="material-icons">accessibility</i>${
         event.attendees
       } Going to Attend</p>
       <p id="modal-details" class="modal-info">${event.details}</p>
@@ -509,22 +522,24 @@ function viewEvent(event) {
 }
 
 function deleteEvent(eventId) {
-  fetch(`/events/${eventId}`, {
-    method: "DELETE",
-  })
-    .then((response) => {
-      if (response.ok) {
-        M.toast({ html: "Event deleted successfully" });
-        location.reload();
-      } else {
-        const modalInstance = M.Modal.getInstance(modal);
-        modalInstance.close();
-        M.toast({ html: "Failed to delete the event" });
-      }
+  if (confirm("Are you sure you want to delete this event?")) {
+    fetch(`/events/${eventId}`, {
+      method: "DELETE",
     })
-    .catch((error) => {
-      console.error("Error:", error);
-    });
+      .then((response) => {
+        if (response.ok) {
+          M.toast({ html: "Event deleted successfully" });
+          location.reload();
+        } else {
+          const modalInstance = M.Modal.getInstance(modal);
+          modalInstance.close();
+          M.toast({ html: "Failed to delete the event" });
+        }
+      })
+      .catch((error) => {
+        console.error("Error:", error);
+      });
+  }
 }
 
 function updateEvent(eventId, updatedEvent) {
@@ -590,18 +605,20 @@ function updateRSVP(id) {
 }
 
 async function deleteCategory(id) {
-  try {
-    await fetch(`/categories/${id}`, {
-      method: "DELETE",
-    })
-    fetchCategories();
-    M.toast({ html: "Category Deleted" });
-    const row = document.getElementById(`category-row-${id}`);
-    if (row) {
-      row.remove();
+  if (confirm("Are you sure you want to delete this item?")) {
+    try {
+      await fetch(`/categories/${id}`, {
+        method: "DELETE",
+      });
+      fetchCategories();
+      M.toast({ html: "Category Deleted" });
+      const row = document.getElementById(`category-row-${id}`);
+      if (row) {
+        row.remove();
+      }
+    } catch (err) {
+      console.log();
     }
-  } catch (err) {
-    console.log();
   }
 }
